requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
pathlib==1.0.1
pillow==10.4.0
pycryptodome==3.20.0
requests==2.32.3
six==1.16.0
tenacity==5.1.5
tkinter-tooltip==3.1.0
urllib3==2.2.3
